+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Testnet.hs
@@ -16,7 +16,7 @@ import           Control.Monad.Class.MonadFork
 import           Control.Monad.Class.MonadTime.SI (DiffTime, Time (Time),
                      addTime, diffTime)
 import           Control.Monad.IOSim
-import           Data.Bifoldable (bifoldMap)
+import           Data.Bifoldable (bifoldMap, bitraverse_)
 
 import           Data.List (find, intercalate, tails)
 import qualified Data.List.Trace as Trace
@@ -66,6 +66,8 @@ import           Test.QuickCheck
 import           Test.Tasty
 import           Test.Tasty.QuickCheck (testProperty)
 
+import           Control.Exception (AssertionFailed (..), catch, evaluate)
+import           Data.Foldable (foldl')
 import           Ouroboros.Network.BlockFetch (TraceFetchClientState (..))
 import           Ouroboros.Network.ConnectionManager.Test.Timeouts
                      (AllProperty (..), TestProperty (..), classifyActivityType,
@@ -103,6 +105,8 @@ tests =
                    prop_diffusionScript_commandScript_valid
     ]
 #if !defined(mingw32_HOST_OS)
+  , testProperty "no fail"
+                 prop_diffusion_nofail
   , testProperty "no livelock"
                  prop_diffusion_nolivelock
   , testProperty "dns can recover from fails"
@@ -638,6 +642,52 @@ prop_peer_selection_trace_coverage defaultBearerInfo diffScript =
    -- TODO: Add checkCoverage here
    in tabulate "peer selection trace" eventsSeenNames
       True
+
+-- | A variant of 'Test.Ouroboros.Network.PeerSelection.prop_governor_nofail'
+-- but for running on Diffusion.
+--
+prop_diffusion_nofail :: AbsBearerInfo
+                      -> DiffusionScript
+                      -> Property
+prop_diffusion_nofail defaultBearerInfo diffScript =
+    let sim :: SimTrace Void
+        sim = runSimTrace
+            $ diffusionSimulation (toBearerInfo defaultBearerInfo)
+                                  diffScript
+                                  iosimTracer
+
+        trace :: [(Time, DiffusionTestTrace)]
+        trace = Trace.toList
+              . fmap (\(WithTime t (WithName _ b)) -> (t, b))
+              . withTimeNameTraceEvents
+                 @DiffusionTestTrace
+                 @NtNAddr
+              . Trace.fromList (MainReturn (Time 0) () [])
+              . fmap (\(t, tid, tl, te) -> SimEvent t tid tl te)
+              . take 125000
+              . traceEvents
+              $ sim
+
+        -- lastTime :: Time
+        -- lastTime = fst (last trace)
+
+     -- run in `IO` so we can catch the pure 'AssertionFailed' exception
+     in -- classifySimulatedTime lastTime $
+        ioProperty $ do
+       r <-
+         evaluate ( foldl' (flip seq) True
+                $ [ assertPeerSelectionState st ()
+                  | (_, DiffusionDebugPeerSelectionTrace (TraceGovernorState _ _ st)) <- trace ]
+                )
+         `catch` \(AssertionFailed _) -> return False
+       case r of
+         True  -> return $ property True
+         False -> do
+           bitraverse_ (putStrLn . show)
+                       (putStrLn . ppSimEvent 20 20 20)
+                       sim
+           -- the ioSimTrace is infinite, but it will terminate with `AssertionFailed`
+           error "impossible!"
 
 -- | A variant of
 -- 'Test.Ouroboros.Network.ConnectionHandler.Network.PeerSelection.prop_governor_nolivelock'
@@ -1357,6 +1407,12 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
               Governor.inProgressPromoteCold
               events
 
+          govInProgressDemoteToColdSig :: Signal (Set NtNAddr)
+          govInProgressDemoteToColdSig =
+            selectDiffusionPeerSelectionState
+              Governor.inProgressDemoteToCold
+              events
+
           govEstablishedPeersSig :: Signal (Set NtNAddr)
           govEstablishedPeersSig =
             selectDiffusionPeerSelectionState
@@ -1424,7 +1480,7 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
 
           promotionOpportunities :: Signal (Set NtNAddr)
           promotionOpportunities =
-            (\local established recentFailures inProgressPromoteCold isAlive ->
+            (\local established recentFailures inProgressPromoteCold isAlive inProgressDemoteToCold ->
               if isAlive then
                 Set.unions
                   [ -- There are no opportunities if we're at or above target
@@ -1433,6 +1489,7 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
                        else groupEstablished Set.\\ established
                                              Set.\\ recentFailures
                                              Set.\\ inProgressPromoteCold
+                                             Set.\\ inProgressDemoteToCold
                   | (_, WarmValency warmTarget, group) <- LocalRootPeers.toGroupSets local
                   , let groupEstablished = group `Set.intersection` established
                   ]
@@ -1442,6 +1499,7 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
               <*> govEstablishedFailuresSig
               <*> govInProgressPromoteColdSig
               <*> trIsNodeAlive
+              <*> govInProgressDemoteToColdSig
 
           promotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           promotionOpportunitiesIgnoredTooLong =
@@ -1525,6 +1583,12 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
           govActiveTargetsSig =
             selectDiffusionPeerSelectionState
               (targetNumberOfActivePeers . Governor.targets)
+              events
+
+          govInProgressDemoteToColdSig :: Signal (Set NtNAddr)
+          govInProgressDemoteToColdSig =
+            selectDiffusionPeerSelectionState
+              Governor.inProgressDemoteToCold
               events
 
           govEstablishedPeersSig :: Signal (Set NtNAddr)
@@ -1617,11 +1681,12 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
           -- want to promote any, since we'd then be above target for the local
           -- root peer group.
           --
-          promotionOpportunity target local established active recentFailures isAlive
+          promotionOpportunity target local established active recentFailures isAlive inProgressDemoteToCold
             | isAlive && Set.size active < target
             = established Set.\\ active
                           Set.\\ LocalRootPeers.keysSet local
                           Set.\\ recentFailures
+                          Set.\\ inProgressDemoteToCold
 
             | otherwise
             = Set.empty
@@ -1635,6 +1700,7 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
               <*> govActivePeersSig
               <*> govActiveFailuresSig
               <*> trIsNodeAlive
+              <*> govInProgressDemoteToColdSig
 
           promotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           promotionOpportunitiesIgnoredTooLong =
@@ -1716,6 +1782,12 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
               (EstablishedPeers.toSet . Governor.establishedPeers)
               events
 
+          govInProgressDemoteToColdSig :: Signal (Set NtNAddr)
+          govInProgressDemoteToColdSig =
+            selectDiffusionPeerSelectionState
+              Governor.inProgressDemoteToCold
+              events
+
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
               Signal.fromChangeEvents Killed -- Default to TrKillingNode
@@ -1778,7 +1850,7 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
 
           promotionOpportunities :: Signal (Set NtNAddr)
           promotionOpportunities =
-            (\local established active recentFailures isAlive ->
+            (\local established active recentFailures isAlive inProgressDemoteToCold ->
               if isAlive then
                 Set.unions
                   [ -- There are no opportunities if we're at or above target
@@ -1786,6 +1858,7 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
                        then Set.empty
                        else groupEstablished Set.\\ active
                                              Set.\\ recentFailures
+                                             Set.\\ inProgressDemoteToCold
                   | (HotValency hotTarget, _, group) <- LocalRootPeers.toGroupSets local
                   , let groupActive      = group `Set.intersection` active
                         groupEstablished = group `Set.intersection` established
@@ -1797,6 +1870,7 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
               <*> govActivePeersSig
               <*> govActiveFailuresSig
               <*> trIsNodeAlive
+              <*> govInProgressDemoteToColdSig
 
           promotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           promotionOpportunitiesIgnoredTooLong =
@@ -2098,6 +2172,10 @@ prop_diffusion_target_active_local_above defaultBearerInfo diffScript =
           govActivePeersSig =
             selectDiffusionPeerSelectionState Governor.activePeers events
 
+          govInProgressDemoteToColdSig :: Signal (Set NtNAddr)
+          govInProgressDemoteToColdSig =
+            selectDiffusionPeerSelectionState Governor.inProgressDemoteToCold events
+
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
               Signal.fromChangeEvents Killed -- Default to TrKillingNode
@@ -2127,13 +2205,14 @@ prop_diffusion_target_active_local_above defaultBearerInfo diffScript =
 
           demotionOpportunities :: Signal (Set NtNAddr)
           demotionOpportunities =
-            (\local active isAlive ->
+            (\local active isAlive inProgressDemoteToCold ->
               if isAlive
               then Set.unions
                     [ -- There are no opportunities if we're at or below target
                       if Set.size groupActive <= hotTarget
                          then Set.empty
                          else groupActive
+                                Set.\\ inProgressDemoteToCold
                     | (HotValency hotTarget, _, group) <- LocalRootPeers.toGroupSets local
                     , let groupActive = group `Set.intersection` active
                     ]
@@ -2141,6 +2220,7 @@ prop_diffusion_target_active_local_above defaultBearerInfo diffScript =
             ) <$> govLocalRootPeersSig
               <*> govActivePeersSig
               <*> trIsNodeAlive
+              <*> govInProgressDemoteToColdSig
 
           demotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           demotionOpportunitiesIgnoredTooLong =
@@ -2617,9 +2697,20 @@ prop_diffusion_peer_selection_actions_no_dodgy_traces defaultBearerInfo (HotDiff
                f as = conjoin $ g <$> tails as
 
                g :: [WithTime (PeerSelectionActionsTrace NtNAddr NtNVersion)] -> Property
-               g as@(WithTime demotionTime (PeerStatusChanged HotToWarm{}) : as') =
+               g as@(WithTime demotionTime (PeerStatusChanged HotToCooling{}) : as') =
                  case find (\case
-                               WithTime _ (PeerStatusChanged WarmToHot{}) -> True
+                               WithTime _ (PeerStatusChanged ColdToWarm{}) -> True
+                               _ -> False)
+                           as' of
+
+                   Nothing                         -> property True
+                   Just (WithTime promotionTime _) -> counterexample (show as)
+                                                    $ ( promotionTime `diffTime` demotionTime
+                                                     >= reconnectDelay config_RECONNECT_DELAY
+                                                      )
+               g as@(WithTime demotionTime (PeerStatusChanged WarmToCooling{}) : as') =
+                 case find (\case
+                               WithTime _ (PeerStatusChanged ColdToWarm{}) -> True
                                _ -> False)
                            as' of