+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor.hs
@@ -49,8 +49,6 @@ import           Control.Monad.Class.MonadTimer.SI
 import           Control.Tracer (Tracer (..), traceWith)
 import           System.Random
 
-import           Data.Set (Set)
-import qualified Data.Set as Set
 import           Ouroboros.Network.PeerSelection.Churn (peerChurnGovernor)
 import qualified Ouroboros.Network.PeerSelection.Governor.ActivePeers as ActivePeers
 import qualified Ouroboros.Network.PeerSelection.Governor.BigLedgerPeers as BigLedgerPeers
@@ -507,13 +505,11 @@ peerSelectionGovernorLoop tracer
                           policy
                           jobPool
                           pst = do
-    cacheTVar <- newTVarIO Set.empty
-    loop cacheTVar pst
+    loop pst
   where
-    loop :: StrictTVar m (Set peeraddr)
-         -> PeerSelectionState peeraddr peerconn
+    loop :: PeerSelectionState peeraddr peerconn
          -> m Void
-    loop cacheTVar !st = assertPeerSelectionState st $ do
+    loop !st = assertPeerSelectionState st $ do
       -- Update public state using 'toPublicState' to compute available peers
       -- to share for peer sharing
       atomically $ writeTVar stateVar (toPublicState st)
@@ -523,7 +519,7 @@ peerSelectionGovernorLoop tracer
           st'               = st { knownPeers       = knownPeers',
                                    establishedPeers = establishedPeers' }
 
-      timedDecision <- evalGuardedDecisions blockedAt (peerSharing actions) cacheTVar st'
+      timedDecision <- evalGuardedDecisions blockedAt (peerSharing actions) st'
 
       -- get the current time after the governor returned from the blocking
       -- 'evalGuardedDecisions' call.
@@ -537,15 +533,14 @@ peerSelectionGovernorLoop tracer
                      newCounters
 
       mapM_ (JobPool.forkJob jobPool) decisionJobs
-      loop cacheTVar (decisionState { countersCache = Cache newCounters })
+      loop (decisionState { countersCache = Cache newCounters })
 
     evalGuardedDecisions :: Time
                          -> PeerSharing
-                         -> StrictTVar m (Set peeraddr)
                          -> PeerSelectionState peeraddr peerconn
                          -> m (TimedDecision m peeraddr peerconn)
-    evalGuardedDecisions blockedAt peerSharing cacheTVar st =
-      case guardedDecisions blockedAt peerSharing cacheTVar st of
+    evalGuardedDecisions blockedAt peerSharing st =
+      case guardedDecisions blockedAt peerSharing st of
         GuardedSkip _ ->
           -- impossible since guardedDecisions always has something to wait for
           error "peerSelectionGovernorLoop: impossible: nothing to do"
@@ -567,12 +562,11 @@ peerSelectionGovernorLoop tracer
 
     guardedDecisions :: Time
                      -> PeerSharing
-                     -> StrictTVar m (Set peeraddr)
                      -> PeerSelectionState peeraddr peerconn
                      -> Guarded (STM m) (TimedDecision m peeraddr peerconn)
-    guardedDecisions blockedAt peerSharing cacheTVar st =
+    guardedDecisions blockedAt peerSharing st =
       -- All the alternative potentially-blocking decisions.
-         Monitor.connections          cacheTVar actions st
+         Monitor.connections          actions st
       <> Monitor.jobs                 jobPool st
       <> Monitor.targetPeers          actions st
       <> Monitor.localRoots           actions policy st